requirements.txt
adafruit-circuitpython-dht==4.0.9
azure-iot-device==2.14.0
dotenv==0.9.9